Move moduleProps.tsx into _lib folder
